code reorganization: extracted PRTContext into separate files
